prefect_flow/requirements.txt
optuna==3.6.1
numpy==1.26.4
Pillow==10.4.0
pandas==2.2.2
SQLAlchemy==2.0.32
tensorflow-cpu==2.16.1
scikit-learn==1.5.1
mlflow==2.15.1
requests==2.32.3
loguru==0.7.2
